requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
tiktoken
beautifulsoup4
lxml
scikit-learn
pandas
pdfplumber==0.11.6
pdfminer.six==20250327
python-docx
openpyxl
fastapi
motor
pymongo
python-multipart
faiss-cpu
numpy==1.26.4
